Link to example of incremental BLOB I/O usage
I find the test code very helpful for understanding how to use the API, but every time I'm away from this library for a few months, I forget how to find it and have to spend a few minutes digging around for it. I hope that by linking directly to the tests, it makes it easier for clients to understand how to use the incremental BLOB I/O feature.